Bug fixes - Fixed all warnings `-Wall` `-Wextra` `-Wpedantic` - Fixed a bug with `editor_new` sourced files

+++ buf.h
@@ -33,7 +33,7 @@ void buf_free(struct buf *);
  * tgt - The buffer to read into
  * Returns: The number of bytes read
  */
-int buf_read(struct buf const *, int pos, int cnt, char *tgt);
+size_t buf_read(struct buf const *, size_t pos, size_t cnt, char *tgt);
 
 /*
  * Writes a set number of bytes to the buffer
@@ -42,11 +42,11 @@ int buf_read(struct buf const *, int pos, int cnt, char *tgt);
  * cnt - The number of bytes to write
  * tgt - The buffer to write from
  */
-void buf_write(struct buf **, int pos, int cnt, char const *tgt);
+void buf_write(struct buf **, size_t pos, size_t cnt, char const *tgt);
 
 /*
  * Get the current size of the buffer
  */
-int buf_sz(struct buf const *);
+size_t buf_sz(struct buf const *);
 
 #endif // BUF_H_
